ec_csup.cpp is 100% matched!

+++ src/sdk/ec/csup.h
@@ -72,5 +72,5 @@ struct ECProgress {
 };
 
 namespace ec {
-    unsigned long long getExternalErrCode(ECProgress *progress);
+    unsigned long getExternalErrCode(ECProgress *progress);
 }
